Add some changes
This adds the "Testing Phase" to keep everything that is related to
quality assurance (QA) before actually releasing the software. After it
has been released the maintenance phase is in effect until the next
feature is planned.

Additionally the maintenance section was written and open tasks where
added. The continous processes have been moved to the application
lifecyle chapter, as these affect each stage.

A figure depicting the application development lifecyle was added. It
was done in TeX because that was the quickest way to go – no need to
keep it that way.